requirements.txt
streamlit
beautifulsoup4
requests
python-dotenv
linkedin-api
openai
selenium
chromedriver